requirements.txt
flask
pandas
scikit-learn
mysql-connector-python
